HS-1350 small refactoring
