+++ metrics-processor/main/src/test/java/org/opennms/horizon/flows/FlowProcessorTest.java
@@ -28,6 +28,7 @@
 
 package org.opennms.horizon.flows;
 
+import org.junit.Ignore;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
