--- metrics-processor/main/src/test/java/org/opennms/horizon/flows/KafkaProducer.java
@@ -1,82 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2023 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2023 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.horizon.flows;
-
-import java.nio.charset.StandardCharsets;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.common.header.Header;
-import org.apache.kafka.common.header.internals.RecordHeader;
-import org.junit.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.support.SendResult;
-import org.springframework.stereotype.Component;
-
-@Component
-@TestConfiguration
-public class KafkaProducer {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(KafkaProducer.class);
-
-    @Value("${kafka.flow-topics}")
-    private String flowTopic;
-
-    @Autowired
-    @Qualifier("kafkaByteArrayProducerTemplate")
-    private KafkaTemplate<String, byte[]> kafkaTemplate;
-
-
-    public CompletableFuture<SendResult<String, byte[]>> send(String topic, byte[] payload, String tenantId) {
-        ProducerRecord<String, byte[]> producerRecord = formatProducerRecord(payload, tenantId);
-        LOGGER.info("sending payload='{}' to topic='{}'", producerRecord, topic);
-        return kafkaTemplate.send(producerRecord);
-    }
-
-    private ProducerRecord<String, byte[]> formatProducerRecord(byte[] rawContent, String tenantId) {
-        List<Header> headers = new LinkedList<>();
-        headers.add(new RecordHeader(tenantId, tenantId.getBytes(StandardCharsets.UTF_8)));
-
-        return new ProducerRecord<String, byte[]>(
-            flowTopic,
-            null,
-            null,
-            rawContent,
-            headers
-        );
-    }
-}
