+++ metrics-processor/main/src/test/java/org/opennms/horizon/flows/grpc/client/IngestorApplicationConfig.java
@@ -33,6 +33,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
+import org.springframework.retry.backoff.FixedBackOffPolicy;
+import org.springframework.retry.policy.SimpleRetryPolicy;
+import org.springframework.retry.support.RetryTemplate;
 
 import io.grpc.ManagedChannel;
 import io.grpc.inprocess.InProcessChannelBuilder;
@@ -40,7 +43,7 @@ import io.grpc.inprocess.InProcessServerBuilder;
 
 @TestConfiguration
 @ImportAutoConfiguration
-public class InventoryApplicationConfigTest {
+public class IngestorApplicationConfig {
 
     public static final String SERVER_NAME = InProcessServerBuilder.generateName();
 
@@ -48,18 +51,23 @@ public class InventoryApplicationConfigTest {
     private long deadline;
 
     @Bean
-    public GrpcInventoryMockServer grpcInventoryMockServer() {
-        return new GrpcInventoryMockServer();
+    public GrpcIngesterMockServer grpcIngesterMockServer() {
+        return new GrpcIngesterMockServer();
     }
 
-    @Bean(name = "inventoryChannel")
-    public ManagedChannel createInventoryChannel() {
+    @Bean(name = "ingestorChannel")
+    public ManagedChannel createIngestorChannel() {
         return InProcessChannelBuilder.forName(SERVER_NAME).directExecutor().build();
     }
 
-    @Bean(destroyMethod = "shutdown", initMethod = "initialStubs")
-    public InventoryClient createInventoryClient(@Qualifier("inventoryChannel") ManagedChannel channel) {
-        return new InventoryClient(channel, deadline);
+    @Bean(destroyMethod = "shutdown", initMethod = "initStubs")
+    public IngestorClient createIngestorClient(@Qualifier("ingestorChannel") ManagedChannel channel, RetryTemplate retryTemplate) {
+        return new IngestorClient(channel, deadline, retryTemplate);
+    }
+
+    @Bean
+    public RetryTemplate retryTemplate() {
+        return new RetryTemplate();
     }
 
 }
