+++ metrics-processor/main/src/test/java/org/opennms/horizon/metrics/FlowProcessorIntegrationTest.java
@@ -1,14 +1,20 @@
-package org.opennms.horizon.flows;
+package org.opennms.horizon.metrics;
 
 
 import static org.awaitility.Awaitility.await;
 
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.time.Duration;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
+import org.apache.kafka.clients.producer.ProducerRecord;
+import org.apache.kafka.common.header.Header;
+import org.apache.kafka.common.header.internals.RecordHeader;
 import org.junit.Rule;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
@@ -16,20 +22,21 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.opennms.dataplatform.flows.document.FlowDocument;
 import org.opennms.dataplatform.flows.document.FlowDocumentLog;
+import org.opennms.horizon.flows.FlowProcessorTestConfig;
 import org.opennms.horizon.flows.grpc.client.GrpcIngesterMockServer;
 import org.opennms.horizon.flows.grpc.client.GrpcInventoryMockServer;
-import org.opennms.horizon.flows.grpc.client.IngestorApplicationConfigTest;
-import org.opennms.horizon.flows.grpc.client.InventoryApplicationConfigTest;
+import org.opennms.horizon.flows.grpc.client.IngestorApplicationConfig;
+import org.opennms.horizon.flows.grpc.client.InventoryApplicationConfig;
 import org.opennms.horizon.inventory.dto.NodeIdQuery;
-import org.opennms.horizon.metrics.MetricsProcessorApplication;
 import org.opennms.taskset.contract.TaskResult;
 import org.opennms.taskset.contract.TaskSetResults;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.support.SendResult;
 import org.springframework.kafka.test.context.EmbeddedKafka;
 import org.springframework.test.annotation.DirtiesContext;
@@ -42,17 +49,13 @@ import io.grpc.inprocess.InProcessServerBuilder;
 import io.grpc.testing.GrpcCleanupRule;
 
 @SpringBootTest
-@ContextConfiguration(classes = {KafkaProducer.class, KafkaTestConfig.class, MetricsProcessorApplication.class, IngestorApplicationConfigTest.class,
-    InventoryApplicationConfigTest.class, FlowProcessorTestConfig.class})
-@EmbeddedKafka(brokerProperties = {"listeners=PLAINTEXT://localhost:9092", "port=9092"}, topics = "flows")
+@ContextConfiguration(classes = {IngestorApplicationConfig.class, InventoryApplicationConfig.class,
+    FlowProcessorTestConfig.class})
+@EmbeddedKafka(brokerProperties = {"listeners=PLAINTEXT://localhost:59092", "port=59092"}, topics = "flows-test")
 @DirtiesContext
-@EnableConfigurationProperties
 @TestPropertySource(locations = "/application-test.yml")
 @ActiveProfiles("test")
 class FlowProcessorIntegrationTest {
-
-    @Autowired
-    private KafkaProducer producer;
 
     @Rule
     public final GrpcCleanupRule grpcCleanupRule = new GrpcCleanupRule();
@@ -66,6 +69,10 @@ class FlowProcessorIntegrationTest {
     @Value("${kafka.flow-topics}")
     private String flowTopic;
 
+    @Autowired
+    @Qualifier("kafkaByteArrayProducerTemplate")
+    private KafkaTemplate<String, byte[]> kafkaTemplate;
+
     private static final Logger LOG = LoggerFactory.getLogger(FlowProcessorIntegrationTest.class);
 
     private static final String LOCATION1 = "test-location1";
@@ -77,13 +84,13 @@ class FlowProcessorIntegrationTest {
 
     @BeforeEach
     public void setUpMockServers() throws IOException {
-        ingestorServer = InProcessServerBuilder.forName(IngestorApplicationConfigTest.SERVER_NAME)
+        ingestorServer = InProcessServerBuilder.forName(IngestorApplicationConfig.SERVER_NAME)
             .directExecutor()
             .addService(grpcIngesterMockServer)
             .build().start();
         grpcCleanupRule.register(ingestorServer);
 
-        inventoryServer = InProcessServerBuilder.forName(InventoryApplicationConfigTest.SERVER_NAME)
+        inventoryServer = InProcessServerBuilder.forName(InventoryApplicationConfig.SERVER_NAME)
             .directExecutor()
             .addService(grpcInventoryMockServer)
             .build().start();
@@ -120,8 +127,12 @@ class FlowProcessorIntegrationTest {
             .setExporterAddress("127.0.0.1")
         ).build();
 
+        ProducerRecord<String, byte[]> producerRecord = formatProducerRecord(flows.toByteArray(), tenantId);
+        LOG.info("sending payload='{}' to topic='{}'", producerRecord, flowTopic);
+
+
         // When
-        CompletableFuture<SendResult<String, byte[]>> future = producer.send(flowTopic, flows.toByteArray(), tenantId);
+        CompletableFuture<SendResult<String, byte[]>> future = kafkaTemplate.send(producerRecord);
         future.whenComplete((result, ex) -> Optional.ofNullable(ex)
             .ifPresentOrElse(val -> LOG.error("Unable to send message=[{}] due to: {}", testTaskSetResults.toByteArray(), val.getMessage()),
                 () -> LOG.info("Sent message=[{}] with offset=[{}]", testTaskSetResults.toByteArray(), result.getRecordMetadata().offset())));
@@ -139,5 +150,18 @@ class FlowProcessorIntegrationTest {
         Assertions.assertEquals(2, incomingNodeIdQueries.size());
         Assertions.assertNotNull(incomingNodeIdQueries.get(LOCATION1));
         Assertions.assertNotNull(incomingNodeIdQueries.get(LOCATION2));
+    }
+
+    private ProducerRecord<String, byte[]> formatProducerRecord(byte[] rawContent, String tenantId) {
+        List<Header> headers = new LinkedList<>();
+        headers.add(new RecordHeader(tenantId, tenantId.getBytes(StandardCharsets.UTF_8)));
+
+        return new ProducerRecord<String, byte[]>(
+            flowTopic,
+            null,
+            null,
+            rawContent,
+            headers
+        );
     }
 }
