+++ metrics-processor/main/src/test/resources/application-test.yml
@@ -5,11 +5,11 @@ spring:
       auto-offset-reset: earliest
       key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
       value-deserializer: org.apache.kafka.common.serialization.ByteArrayDeserializer
-    bootstrap-servers: localhost:9092
+    bootstrap-servers: localhost:59092
 
 kafka:
   topics: "task-set.results"
-  flow-topics: "flows"
+  flow-topics: "flows-test"
 
 grpc:
   server:
@@ -22,3 +22,17 @@ grpc:
     retry:
       maxAttempts: 3
       maxDelay: 5000
+
+
+cortex.write.url: http://cortex:9000/api/v1/push
+cortex.maxconcurrenthttpconnections: 100
+cortex.cortexwritetimeoutinms: 1000
+cortex.readtimeoutinms: 1000
+cortex.bulkheadmaxwaitdurationinms: 9223372036854775807
+cortex.organizationid:
+
+management:
+  endpoints:
+    web:
+      exposure:
+        include: "*"
